merge

+++ Common/Product/SharedProject/ProcessOutput.cs
@@ -1,16 +1,18 @@
-﻿/* ****************************************************************************
- *
- * Copyright (c) Microsoft Corporation. 
- *
- * This source code is subject to terms and conditions of the Apache License, Version 2.0. A 
- * copy of the license can be found in the License.html file at the root of this distribution. If 
- * you cannot locate the Apache License, Version 2.0, please send an email to 
- * vspython@microsoft.com. By using this source code in any fashion, you are agreeing to be bound 
- * by the terms of the Apache License, Version 2.0.
- *
- * You must not remove this notice, or any other, from this software.
- *
- * ***************************************************************************/
+﻿// Visual Studio Shared Project
+// Copyright(c) Microsoft Corporation
+// All rights reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the License); you may not use
+// this file except in compliance with the License. You may obtain a copy of the
+// License at http://www.apache.org/licenses/LICENSE-2.0
+//
+// THIS CODE IS PROVIDED ON AN  *AS IS* BASIS, WITHOUT WARRANTIES OR CONDITIONS
+// OF ANY KIND, EITHER EXPRESS OR IMPLIED, INCLUDING WITHOUT LIMITATION ANY
+// IMPLIED WARRANTIES OR CONDITIONS OF TITLE, FITNESS FOR A PARTICULAR PURPOSE,
+// MERCHANTABLITY OR NON-INFRINGEMENT.
+//
+// See the Apache Version 2.0 License for specific language governing
+// permissions and limitations under the License.
 
 using System;
 using System.Collections.Generic;
@@ -24,7 +26,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.Win32.SafeHandles;
 
-namespace Microsoft.VisualStudioTools.Project {
+namespace Microsoft.VisualStudioTools.Infrastructure {
     /// <summary>
     /// Base class that can receive output from <see cref="ProcessOutput"/>.
     /// 
@@ -200,6 +202,7 @@ namespace Microsoft.VisualStudioTools.Project {
             psi.UseShellExecute = false;
             psi.RedirectStandardError = !visible || (redirector != null);
             psi.RedirectStandardOutput = !visible || (redirector != null);
+            psi.RedirectStandardInput = !visible;
             psi.StandardOutputEncoding = outputEncoding ?? psi.StandardOutputEncoding;
             psi.StandardErrorEncoding = errorEncoding ?? outputEncoding ?? psi.StandardErrorEncoding;
             if (env != null) {
@@ -422,6 +425,15 @@ namespace Microsoft.VisualStudioTools.Project {
                 }
                 if (_process.StartInfo.RedirectStandardError) {
                     _process.BeginErrorReadLine();
+                }
+                
+                if (_process.StartInfo.RedirectStandardInput) {
+                    // Close standard input so that we don't get stuck trying to read input from the user.
+                    try {
+                        _process.StandardInput.Close();
+                    } catch (InvalidOperationException) {
+                        // StandardInput not available
+                    }
                 }
             }
         }
@@ -678,7 +690,7 @@ namespace Microsoft.VisualStudioTools.Project {
         /// Immediately stops the process.
         /// </summary>
         public void Kill() {
-            if (_process != null) {
+            if (_process != null && !_process.HasExited) {
                 _process.Kill();
                 // Should have already been called, in which case this is a no-op
                 OnExited(this, EventArgs.Empty);
